nocoloco/requirements.txt
nornir==3.3.0
nornir-utils==0.2.0
nornir-jinja2==0.2.0
nornir-napalm==0.4.0
nornir-salt==0.20.2
nornir-netmiko==1.0.0
ttp==0.9.5
genie==23.9
pyats==23.9
tabulate==0.9.0
Flask[async]==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.8.6
requests==2.28.2
nornir-sql==0.0.4.post1
waitress==2.1.2
werkzeug==2.2.3
python-dotenv==1.0.0